Further refactoring to QOpenGLWidget.
