+++ src/opencmiss/zincwidgets/sceneviewerwidget.py
@@ -205,7 +205,7 @@ class SceneviewerWidget(QtWidgets.QOpenGLWidget):
         repaint required event all other events are ignored.
         """
         if event.getChangeFlags() & Sceneviewerevent.CHANGE_FLAG_REPAINT_REQUIRED:
-            QtCore.QTimer.singleShot(0, self.updateGL)
+            QtCore.QTimer.singleShot(0, self.update)
 
 #  Not applicable at the current point in time.
 #     def _zincSelectionEvent(self, event):
