+++ src/opencmiss/zincwidgets/spectrumeditorwidget.py
@@ -29,10 +29,10 @@ SPECTRUM_DATA_ROLE = QtCore.Qt.UserRole + 1
 
 class SpectrumEditorWidget(QtWidgets.QWidget):
 
-    def __init__(self, parent=None, shared_context=None):
+    def __init__(self, parent=None):
         super(SpectrumEditorWidget, self).__init__(parent)
         self._ui = Ui_SpectrumEditorWidget()
-        self._ui.setupUi(shared_context,self)
+        self._ui.setupUi(self)
 
         self._ui.comboBoxColourMap.addItems(extractColourMappingEnum())
         self._ui.comboBoxScale.addItems(extractScaleTypeEnum())
